add search

+++ weblog-module-admin/src/main/java/com/mll/weblog/admin/service/impl/AdminCategoryServiceImpl.java
@@ -1,13 +1,14 @@
 package com.mll.weblog.admin.service.impl;
 
-import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.mll.weblog.admin.model.vo.category.AddCategoryReqVO;
 import com.mll.weblog.admin.model.vo.category.DeleteCategoryReqVO;
 import com.mll.weblog.admin.model.vo.category.FindCategoryPageListReqVO;
 import com.mll.weblog.admin.model.vo.category.FindCategoryPageListRspVO;
 import com.mll.weblog.admin.service.AdminCategoryService;
+import com.mll.weblog.common.domain.dos.ArticleCategoryRelDO;
 import com.mll.weblog.common.domain.dos.CategoryDO;
+import com.mll.weblog.common.domain.mapper.ArticleCategoryRelMapper;
 import com.mll.weblog.common.domain.mapper.CategoryMapper;
 import com.mll.weblog.common.enums.ResponseCodeEnum;
 import com.mll.weblog.common.exception.BizException;
@@ -15,7 +16,6 @@ import com.mll.weblog.common.model.vo.SelectRspVO;
 import com.mll.weblog.common.utils.PageResponse;
 import com.mll.weblog.common.utils.Response;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
@@ -26,18 +26,19 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
- * @Title: AdminCategoryServiceImpl
- * @Author mll
- * @Package com.mll.weblog.admin.service.impl
- * @Date 2024/10/14 16:37
- * @description: 新增分类服务实现类
- */
+ * @author: mll
+ * @url: www.mll.com
+ * @date: 2024-09-15 14:03
+ * @description: TODO
+ **/
 @Service
 @Slf4j
 public class AdminCategoryServiceImpl implements AdminCategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private ArticleCategoryRelMapper articleCategoryRelMapper;
 
     /**
      * 添加分类
@@ -46,14 +47,14 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
      * @return
      */
     @Override
-    public Response<AddCategoryReqVO> addCategory(AddCategoryReqVO addCategoryReqVO) {
+    public Response addCategory(AddCategoryReqVO addCategoryReqVO) {
         String categoryName = addCategoryReqVO.getName();
 
         // 先判断该分类是否已经存在
         CategoryDO categoryDO = categoryMapper.selectByName(categoryName);
 
         if (Objects.nonNull(categoryDO)) {
-            log.warn("分类名称： {}, 此分类已存在", categoryName);
+            log.warn("分类名称： {}, 此已存在", categoryName);
             throw new BizException(ResponseCodeEnum.CATEGORY_NAME_IS_EXISTED);
         }
 
@@ -75,29 +76,16 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
      * @return
      */
     @Override
-    public PageResponse findCategoryList(FindCategoryPageListReqVO findCategoryPageListReqVO) {
+    public PageResponse findCategoryPageList(FindCategoryPageListReqVO findCategoryPageListReqVO) {
         // 获取当前页、以及每页需要展示的数据数量
         Long current = findCategoryPageListReqVO.getCurrent();
         Long size = findCategoryPageListReqVO.getSize();
-
-        // 分页对象(查询第几页、每页多少数据)
-        Page<CategoryDO> page = new Page<>(current, size);
-
-        // 构建查询条件
-        LambdaQueryWrapper<CategoryDO> wrapper = new LambdaQueryWrapper<>();
-
         String name = findCategoryPageListReqVO.getName();
         LocalDate startDate = findCategoryPageListReqVO.getStartDate();
         LocalDate endDate = findCategoryPageListReqVO.getEndDate();
 
-        wrapper
-                .like(StringUtils.isNotBlank(name), CategoryDO::getName, name.trim()) // like 模块查询
-                .ge(Objects.nonNull(startDate), CategoryDO::getCreateTime, startDate) // 大于等于 startDate
-                .le(Objects.nonNull(endDate), CategoryDO::getCreateTime, endDate)  // 小于等于 endDate
-                .orderByDesc(CategoryDO::getCreateTime); // 按创建时间倒叙
-
         // 执行分页查询
-        Page<CategoryDO> categoryDOPage = categoryMapper.selectPage(page, wrapper);
+        Page<CategoryDO> categoryDOPage = categoryMapper.selectPageList(current, size, name, startDate, endDate);
 
         List<CategoryDO> categoryDOS = categoryDOPage.getRecords();
 
@@ -109,6 +97,7 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
                             .id(categoryDO.getId())
                             .name(categoryDO.getName())
                             .createTime(categoryDO.getCreateTime())
+                            .articlesTotal(categoryDO.getArticlesTotal())
                             .build())
                     .collect(Collectors.toList());
         }
@@ -126,6 +115,14 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
     public Response deleteCategory(DeleteCategoryReqVO deleteCategoryReqVO) {
         // 分类 ID
         Long categoryId = deleteCategoryReqVO.getId();
+
+        // 校验该分类下是否已经有文章，若有，则提示需要先删除分类下所有文章，才能删除
+        ArticleCategoryRelDO articleCategoryRelDO = articleCategoryRelMapper.selectOneByCategoryId(categoryId);
+
+        if (Objects.nonNull(articleCategoryRelDO)) {
+            log.warn("==> 此分类下包含文章，无法删除，categoryId: {}", categoryId);
+            throw new BizException(ResponseCodeEnum.CATEGORY_CAN_NOT_DELETE);
+        }
 
         // 删除分类
         categoryMapper.deleteById(categoryId);
@@ -158,4 +155,5 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
 
         return Response.success(selectRspVOS);
     }
+
 }
